// src/lib/api/generated/model/campaignStep.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Django Continuous Integration API
 * API documentation for the Django Continuous Integration project with tenant-aware access rules.
 * OpenAPI spec version: 1.0.0
 */
import type { StepTypeEnum } from './stepTypeEnum';
import type { SendConditionEnum } from './sendConditionEnum';

export interface CampaignStep {
  readonly id: string;
  readonly campaign: string;
  readonly tenant: string;
  step_type?: StepTypeEnum;
  /**
   * @minimum 0
   * @maximum 9223372036854776000
   */
  order?: number;
  email_template: string;
  /**
   * @minimum 0
   * @maximum 9223372036854776000
   */
  delay_days?: number;
  /**
   * @minimum 0
   * @maximum 9223372036854776000
   */
  delay_hours?: number;
  send_condition?: SendConditionEnum;
  readonly created_at: string;
  readonly updated_at: string;
}