// src/lib/api/generated/model/status372Enum.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Django Continuous Integration API
 * API documentation for the Django Continuous Integration project with tenant-aware access rules.
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `ACTIVE` - Active
* `INACTIVE` - Inactive
 */
export type Status372Enum = typeof Status372Enum[keyof typeof Status372Enum];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const Status372Enum = {
  ACTIVE: 'ACTIVE',
  INACTIVE: 'INACTIVE',
} as const;